# tests/snapshots/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunRequest": {
        "properties": {
          "backend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "local",
            "title": "Backend"
          },
          "job_name": {
            "title": "Job Name",
            "type": "string"
          },
          "workspace_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Dir"
          }
        },
        "required": [
          "job_name"
        ],
        "title": "RunRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Data Platform Copilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/build/": {
      "post": {
        "operationId": "build_pipeline_api_v1_build__post",
        "parameters": [
          {
            "in": "query",
            "name": "base_ref",
            "required": false,
            "schema": {
              "default": "upstream/main",
              "title": "Base Ref",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_intelligence",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Intelligence",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Pipeline",
        "tags": [
          "Build"
        ]
      }
    },
...
    "/api/v2/build/": {
      "post": {
        "operationId": "build_pipeline_api_v2_build__post",
        "parameters": [
          {
            "in": "query",
            "name": "base_ref",
            "required": false,
            "schema": {
              "default": "upstream/main",
              "title": "Base Ref",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_intelligence",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Intelligence",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Pipeline",
        "tags": [
          "Build"
        ]
      }
    },
...
    "/api/v2/build/run": {
      "post": {
        "operationId": "run_build_api_v2_build_run_post",
        "parameters": [
          {
            "in": "query",
            "name": "contract_hash",
            "required": true,
            "schema": {
              "title": "Contract Hash",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Run Build Api V2 Build Run Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Build",
        "tags": [
          "execution"
        ]
      }
    },
...
    "/api/v3/build/run": {
      "post": {
        "operationId": "run_build_api_v3_build_run_post",
        "parameters": [
          {
            "in": "query",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "contract_hash",
            "required": true,
            "schema": {
              "title": "Contract Hash",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "plan_hash",
            "required": true,
            "schema": {
              "title": "Plan Hash",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "intelligence_hash",
            "required": true,
            "schema": {
              "title": "Intelligence Hash",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "policy_eval_hash",
            "required": true,
            "schema": {
              "title": "Policy Eval Hash",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Build",
        "tags": [
          "build_v3"
        ]
      }
    },
...
    "/build/": {
      "post": {
        "operationId": "build_pipeline_build__post",
        "parameters": [
          {
            "in": "query",
            "name": "base_ref",
            "required": false,
            "schema": {
              "default": "upstream/main",
              "title": "Base Ref",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_intelligence",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Intelligence",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Pipeline",
        "tags": [
          "Build"
        ]
      }
    },
...
  }
}